Added exits for admin

+++ src/main/java/prabhjot/safin/retail/apps/AdminApp.java
@@ -46,7 +46,7 @@ public class AdminApp extends Application {
             System.out.println("Goodbye !");
             connectionProvider.kill();
         } catch (SQLException e) {
-            System.out.println(e.getMessage());
+            handleSQLException(e.getMessage());
         } catch (InputMismatchException e) {
             System.out.println("Not a valid option !");
             sc.next();
@@ -57,14 +57,19 @@ public class AdminApp extends Application {
         while (true) {
             try {
                 Console console = System.console();
+                
                 System.out.println("Enter your admin id: ");
                 int id = sc.nextInt();
+                
                 System.out.println("Enter your password: ");
                 String password = "";
+                
                 char[] passwordInput = console.readPassword();
                 for (char c : passwordInput) password += c;
+                
                 Admin admin = new Admin(id, password);
                 int status = adminService.login(admin);
+
                 if (status == 0) {
                     System.out.println("Hello admin #" + admin.getId());
                     break;
@@ -93,8 +98,10 @@ public class AdminApp extends Application {
                 System.out.println("Enter 5 to delete a category by id");
                 System.out.println("Enter 6 to exit categories");
                 System.out.println("--------------------------------------");
+                
                 int input = sc.nextInt();
                 sc.nextLine();
+                
                 if      (input == 1) printCategories();
                 else if (input == 2) getCategoryById();
                 else if (input == 3) createCategory();
@@ -121,36 +128,62 @@ public class AdminApp extends Application {
     }
 
     private void getCategoryById() throws ClassNotFoundException, SQLException {
+        showCancelInteger();
         System.out.println("Enter id of category: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         Category category = categoryService.getCategory(id);
         System.out.println(category);
     }
 
     private void createCategory() throws SQLException, ClassNotFoundException {
+        showCancelString();        
         System.out.println("Enter category name: ");
+        
         String name = sc.nextLine();
+        
+        if (cancelStringOperation(name)) return;
+        
         categoryService.createCategory(new Category(name));
         System.out.println("Successfully created category !");
     }
 
     private void updateCategory() throws ClassNotFoundException, SQLException {
         printCategories();
+        showCancelInteger();
         System.out.println("Enter id of category from above to update: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
+        showCancelString();
         System.out.println("Enter new category name: ");
+        
         String name = sc.nextLine();
+        
+        if (cancelStringOperation(name)) return;
+        
         categoryService.updateCategory(id, name);
         System.out.println("Successfully updated category !");
     }
 
     private void deleteCategory() throws ClassNotFoundException, SQLException {
         printCategories();
+        showCancelInteger();
         System.out.println("Enter id of category from above to delete: ");
+        
         int id = sc.nextInt();
+        
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         categoryService.deleteCategory(id);
         System.out.println("Successfully deleted category");
     }
@@ -168,7 +201,7 @@ public class AdminApp extends Application {
                 System.out.println("--------------------------------------");
                 int input = sc.nextInt();
                 sc.nextLine();
-                if      (input == 1) this.printProducts();
+                if      (input == 1) printProducts();
                 else if (input == 2) getProductById();
                 else if (input == 3) createProduct();
                 else if (input == 4) updateProduct();
@@ -187,36 +220,65 @@ public class AdminApp extends Application {
     }
     
     private void createProduct() throws ClassNotFoundException, SQLException {
+        showCancelString();
         System.out.println("Enter name of the product: ");
+        
         String productName = sc.nextLine();
+        
+        if (cancelStringOperation(productName)) return;
+        
         printCategories();
+        showCancelInteger();
         System.out.println("Select a category id from above: ");
+        
         int categoryId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(categoryId)) return;
+        
         productService.createProduct(new Product(productName, categoryId));
         System.out.println("Product created !");
     }
 
     private void updateProduct() throws SQLException, ClassNotFoundException {
         printProducts();
+        showCancelInteger();
         System.out.println("Enter product id from above to update: ");
+        
         int productId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(productId)) return;
+        
+        showCancelString();
         System.out.println("Enter new name of new product: ");
         String productName = sc.nextLine();
+        
+        if (cancelStringOperation(productName)) return;
+        
         printCategories();
+        showCancelInteger();
         System.out.println("Select the new category id from above: ");
+        
         int categoryId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(categoryId)) return;
+
         productService.updateProduct(productId, new Product(productName, categoryId));
         System.out.println("Updated product !");
     }
 
     private void deleteProduct() throws SQLException, ClassNotFoundException {
         printProducts();
+        showCancelInteger();
         System.out.println("Enter product id from above to delete: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         productService.deleteProduct(id);
         System.out.println("Product deleted !");
     }
@@ -260,9 +322,14 @@ public class AdminApp extends Application {
     }
 
     private void getReviewById() throws SQLException, ClassNotFoundException {
+        showCancelInteger();
         System.out.println("Enter id of review: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         Review review = reviewService.getReview(id);
         System.out.println(review);
     }
@@ -317,59 +384,109 @@ public class AdminApp extends Application {
     }
 
     private void createStore() throws SQLException, ClassNotFoundException {
+        showCancelString();
         System.out.println("Please enter store name: ");
+        
         String storeName = sc.nextLine();
+        
+        if (cancelStringOperation(storeName)) return;
+        
         storeService.createStore(new Store(storeName));
         System.out.println("Store has been created");
     }
 
     private void updateStore() throws ClassNotFoundException, SQLException {
         printStores();
+        showCancelInteger();
         System.out.println("Please enter store id to update: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
+        showCancelString();
         System.out.println("Please enter new store name: ");
+        
         String storeName = sc.nextLine();
+        
+        if (cancelStringOperation(storeName)) return;
+        
         storeService.updateStore(id, storeName);
         System.out.println("Store has been updated");
     }
 
     private void deleteStore() throws ClassNotFoundException, SQLException {
         printStores();
+        showCancelInteger();
         System.out.println("Please enter store id to delete: ");
+        
         int id = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         storeService.deleteStore(id);
         System.out.println("Store has been deleted");
     }
 
     private void createPrice() throws SQLException, ClassNotFoundException {
         printProducts();
+        showCancelInteger();
         System.out.println("Please choose product id from above: ");
+        
         int productId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(productId)) return;
+        
         printStores();
+        showCancelInteger();
         System.out.println("Please choose store id from above: ");
+        
         int storeId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(storeId)) return;
+        
+        showCancelInteger();
         System.out.println("Please enter a price: ");
+        
         int price = sc.nextInt();
+        
+        if (cancelIntegerOperation(price)) return;
+        
         storeService.createPrice(storeId, productId, price);
         System.out.println("Price has been inserted !");
     }
 
     private void updatePrice() throws SQLException, ClassNotFoundException {
         printProducts();
+        showCancelInteger();
         System.out.println("Please choose product id from above: ");
+        
         int productId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(productId)) return;
+        
         printStores();
+        showCancelInteger();
         System.out.println("Please choose store id from above: ");
+        
         int storeId = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(storeId)) return;
+        
+        showCancelInteger();
         System.out.println("Please enter a new price: ");
+        
         int price = sc.nextInt();
         sc.nextLine();
+        
+        if (cancelIntegerOperation(price)) return;
+        
         storeService.createPrice(storeId, productId, price);
         System.out.println("Price has been updated !");
     }
@@ -418,69 +535,140 @@ public class AdminApp extends Application {
     }
 
     private void getWarehouseById() throws SQLException, ClassNotFoundException {
+        showCancelInteger();
         System.out.println("Please enter warehouse id: ");
-        int id = sc.nextInt(); sc.nextLine();
+        
+        int id = sc.nextInt(); 
+        sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         Warehouse warehouse = warehouseService.getWarehouse(id);
         System.out.println(warehouse);
     }
 
     private void createWarehouse() throws SQLException, ClassNotFoundException {
+        showCancelString();
         System.out.println("Please enter name of new warehouse: ");
+        
         String name = sc.nextLine();
+        
+        if (cancelStringOperation(name)) return;
+        
+        showCancelString();
         System.out.println("Please enter new address of warehouse: ");
+        
         String address = sc.nextLine();
+        
+        if (cancelStringOperation(address)) return;
+        
         warehouseService.create(new Warehouse(name, address));
         System.out.println("Created warehouse !");
     }
 
     private void updateWarehouse() throws SQLException, ClassNotFoundException {
+        showCancelInteger();
         System.out.println("Please enter warehouse id: ");
+        
         int id = sc.nextInt(); sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
+        showCancelString();
         System.out.println("Please enter new name of warehouse: ");
+        
         String name = sc.nextLine();
+        
+        if (cancelStringOperation(name)) return;
+        
+        showCancelString();
         System.out.println("Please enter address of new warehouse: ");
+    
         String address = sc.nextLine();
+        
+        if (cancelStringOperation(address)) return;
+        
         warehouseService.update(id, new Warehouse(name, address));
-        System.out.println("Created warehouse !");
+        System.out.println("Updated warehouse !");
     }
 
     private void deleteWarehouse() throws SQLException {
+        showCancelInteger();
         System.out.println("Please enter warehouse id: ");
+        
         int id = sc.nextInt(); sc.nextLine();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         warehouseService.delete(id);
         System.out.println("Deleted warehouse !");
     }
 
     private void insertProduct() throws SQLException, ClassNotFoundException {
         printWarehouses();
+        showCancelInteger();
         System.out.println("Choose warehouse id from above: ");
+        
         int warehouseId = sc.nextInt();
+        
+        if (cancelIntegerOperation(warehouseId)) return;
+        
         printProducts();
+        showCancelInteger();
         System.out.println("Choose product id from above: ");
+        
         int productId = sc.nextInt();
+        
+        if (cancelIntegerOperation(productId)) return;
+        
+        showCancelInteger();
         System.out.println("Enter quantity: ");
+        
         int quantity = sc.nextInt();
+        
+        if (cancelIntegerOperation(quantity)) return;
+        
         warehouseService.insertProduct(warehouseId, productId, quantity);
         System.out.println("Stock and product has been inserted !");
     }
 
     private void updateProductStock() throws SQLException, ClassNotFoundException {
         printWarehouses();
+        showCancelInteger();
         System.out.println("Choose warehouse id from above: ");
+        
         int warehouseId = sc.nextInt();
+        
+        if (cancelIntegerOperation(warehouseId)) return;
+        
         printProducts();
+        showCancelInteger();
         System.out.println("Choose product id from above: ");
+        
         int productId = sc.nextInt();
+        
+        if (cancelIntegerOperation(productId)) return;
+        
+        showCancelInteger();
         System.out.println("Enter quantity: ");
+        
         int quantity = sc.nextInt();
+        
+        if (cancelIntegerOperation(quantity)) return;
+        
         warehouseService.insertProduct(warehouseId, productId, quantity);
         System.out.println("Stock has been updated !");
     }
 
     private void getProductStock() throws SQLException, ClassNotFoundException {
         printProducts();
+        showCancelInteger();
         System.out.println("Enter product id from above: ");
+        
         int id = sc.nextInt();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         int stock = warehouseService.getProductStock(id);
         System.out.println("Stock of product: " + stock);
     }
@@ -517,8 +705,13 @@ public class AdminApp extends Application {
     }
 
     private void getCustomerById() throws ClassNotFoundException, SQLException {
+        showCancelInteger();
         System.out.println("Enter id: ");
+
         int id = sc.nextInt();
+        
+        if (cancelIntegerOperation(id)) return;
+        
         Customer customer = customerService.getCustomer(id);
         System.out.println(customer);
     }
